requirements.txt
Flask==3.0.0
python-dotenv==1.0.0
pyTelegramBotAPI==4.14.0
psycopg[binary]==3.1.12
